Set the "best" valid preview size and fpsrange cameraPreferences.
Fixes two bugs I'm seeing on my Nexus 4:

AndroidRuntime: java.lang.RuntimeException: setParameters failed
AndroidRuntime: 	at android.hardware.Camera.native_setParameters(Native Method)
AndroidRuntime: 	at android.hardware.Camera.setParameters(Camera.java:2015)

This bug occurs related to the use of a hardcoded preview size (always
choosing the 4th preview size) and fpsrange (always choosing the last),
which actually contains a bug, resulting in:

E QCameraHWI_Parm: status_t android::QCameraHardwareInterface::setPreviewFpsRange(const android::QCameraParameters &): error: FPS range value not supported

Thanks to this question here on stackoverflow:

https://stackoverflow.com/questions/38279166/how-to-resolve-the-android-camera-setparameters-failed

I found the question itself had the more useful info than the answer in
that it shows how to interate through the reported sizes/ranges to
find the "best one", which is just the one with the biggest area.  If this
is not desired, it could be substituted with another method.

Similarly, the various previewfpsranges are searched through, looking for
the "best" one, which I'm assuming is the one with the highest max fps.

Tested on the nexus 4 (android 7.1.2) and a Shield Tablet (Android 6.0.1)